DDP-5802 - split PutFormAnswerRoute tests into classes: because some tests was affected by results of other tests
