+++ pepper-apis/src/test/java/org/broadinstitute/ddp/route/PutFormAnswersRouteStandalone2Test.java
@@ -2,16 +2,12 @@ package org.broadinstitute.ddp.route;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.Assert.assertEquals;
 
 import java.time.Instant;
-import java.util.Set;
 
 import io.restassured.http.ContentType;
 import org.broadinstitute.ddp.constants.ErrorCodes;
 import org.broadinstitute.ddp.db.TransactionWrapper;
-import org.broadinstitute.ddp.db.dao.ActivityInstanceDao;
-import org.broadinstitute.ddp.db.dao.AuthDao;
 import org.broadinstitute.ddp.db.dao.JdbiRevision;
 import org.broadinstitute.ddp.db.dao.QuestionDao;
 import org.broadinstitute.ddp.db.dto.ActivityInstanceDto;
@@ -20,6 +16,10 @@ import org.broadinstitute.ddp.model.activity.definition.validation.RequiredRuleD
 import org.broadinstitute.ddp.model.activity.revision.RevisionMetadata;
 import org.junit.Test;
 
+/**
+ * The test moved to a separate class because result of other tests affected on it and
+ * it's behaviour was incorrect
+ */
 public class PutFormAnswersRouteStandalone2Test extends PutFormAnswersRouteStandaloneTestAbstract {
 
     @Test
@@ -43,28 +43,5 @@ public class PutFormAnswersRouteStandalone2Test extends PutFormAnswersRouteStand
             RevisionMetadata meta = RevisionMetadata.now(user.getUserId(), "remove required");
             handle.attach(QuestionDao.class).disableRequiredRule(questionId, meta);
         });
-    }
-
-    @Test
-    public void testStudyAdmin_hiddenInstance() {
-        ActivityInstanceDto instanceDto = TransactionWrapper.withTxn(handle -> {
-            ActivityInstanceDto dto = insertNewInstanceAndDeferCleanup(handle, form.getActivityId());
-            assertEquals(1, handle.attach(ActivityInstanceDao.class)
-                    .bulkUpdateIsHiddenByActivityIds(testData.getUserId(), true, Set.of(form.getActivityId())));
-            handle.attach(AuthDao.class).assignStudyAdmin(testData.getUserId(), testData.getStudyId());
-            return dto;
-        });
-
-        try {
-            given().auth().oauth2(token)
-                    .pathParam("instanceGuid", instanceDto.getGuid())
-                    .when().put(urlTemplate).then().assertThat()
-                    .log().all()
-                    .statusCode(200);
-        } finally {
-            TransactionWrapper.useTxn(handle -> {
-                handle.attach(AuthDao.class).removeAdminFromAllStudies(testData.getUserId());
-            });
-        }
     }
 }
